refactor: shared/storage/env -> shared/config/env 이동
