+++ apps/web/src/features/auth/lib/useKakaoSDK.ts
@@ -2,7 +2,7 @@
 
 import { useEffect, useState } from 'react';
 
-import { KAKAO_JS_KEY } from '@/shared/storage';
+import { KAKAO_JS_KEY } from '@/shared/config';
 
 declare global {
   interface Window {
