+++ apps/web/src/instrumentation.ts
@@ -1,9 +1,11 @@
 import { captureRequestError } from '@causw/logger';
 
+import { ENVIRONMENT } from '@/shared/config';
+
 export async function register() {
   if (process.env.NEXT_RUNTIME === 'nodejs') {
     await import('../sentry.server.config');
-    if (process.env.NODE_ENV === 'development') {
+    if (ENVIRONMENT === 'development') {
       const { server } = await import('./app/_mock/server');
       server.listen();
       console.log('Server Mock Started');
