+++ apps/web/src/shared/storage/auth/token-manager.ts
@@ -1,7 +1,6 @@
+import { BASE_URL } from '@/shared/config';
 import { DefaultResponseField } from '@/shared/types';
 import { isMobile, isServer } from '@/shared/utils';
-
-import { BASE_URL } from '../env';
 
 import {
   getClientATK,
