+++ apps/web/src/shared/storage/index.ts
@@ -1,2 +1,1 @@
 export * from './auth';
-export * from './env';
